+++ src/page/button/ToolButton.vue
@@ -43,7 +43,41 @@ export default {
   },
   methods: {
     click(button) {
-      this.$emit("click", button);
+      // 数据表格中选中的行数量
+      const tableSelectLength = this.tableSelect.length;
+
+      // 选中数量验证
+      // selectMin 最小选中
+      if (button.selectMin && button.selectMin > tableSelectLength) {
+        this.$message(`最少选中 ${button.selectMin} 条数据, 已选 ${tableSelectLength} 条`);
+        return;
+      }
+      // selectMax 最多选中
+      if (button.selectMax && button.selectMax < tableSelectLength) {
+        this.$message(`最多选中 ${button.selectMax} 条数据, 已选 ${tableSelectLength} 条`);
+        return;
+      }
+      // selectCount 只能选中
+      if (button.selectCount && button.selectCount != tableSelectLength) {
+        this.$message(`只能选中 ${button.selectCount} 条数据, 已选 ${tableSelectLength} 条`);
+        return;
+      }
+      // 校验通过
+      // 调用 click
+      if (button.click && typeof button.click == 'function') {
+        button.click(this.tableSelect, button);
+      }
+      // confirmClick 点击确认
+      if (button.confirmClick && typeof button.confirmClick == 'function') {
+        this.$confirm(`确认${button.name}吗?`, `${button.name}确认`, {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning',
+        }).then(() => {
+          button.confirmClick(this.tableSelect, button);
+        }).catch(() => {
+        });
+      }
     },
     /**
      * 是否隐藏按钮